+++ frontend/assets/components/redactButtonOnClick.js
