+++ frontend/assets/components/resetButtonOnClick.js
@@ -2,4 +2,5 @@ const resetCallout = () => {
   const emptyString = "";
   document.querySelector("#pii-entities").innerHTML = emptyString;
 };
+
 document.querySelector("#reset").addEventListener("click", resetCallout);
