+++ frontend/assets/iframe.html
@@ -32,7 +32,10 @@
         </div>
         <div class="col-auto">
           <button class="c-btn c-btn--primary" id="detect">Detect PII</button>
-          <script type="module" src="components/predictSegment.js"></script>
+          <script
+            type="module"
+            src="components/piiDetectButtonOnClick.js"
+          ></script>
         </div>
       </div>
     </div>
@@ -49,11 +52,14 @@
       <div class="row justify-content-between">
         <div class="col-auto">
           <button class="c-btn c-btn--primary" id="reset">Reset</button>
-          <script src="components/resetButton.js"></script>
+          <script src="components/resetButtonOnClick.js"></script>
         </div>
         <div class="col-auto">
           <button class="c-btn c-btn--primary" id="redact">Redact</button>
-          <script type="module" src="components/redactSegment.js"></script>
+          <script
+            type="module"
+            src="components/redactButtonOnClick.js"
+          ></script>
         </div>
       </div>
     </div>
